Fixed constraints in VAB/SPH
When the constraint on a part was moved past where the part actually
was, the current code only rewrote the position of the part, but did not
actually move it to the new position.  This commit allows execution of
the rest of the FixedUpdate() inside the VAB/SPH instead of only in
flight.  When it checks limits, if it finds the part is out of limits,
it then moves it and resets the current rotation/translation.  It also
limits maxTweak to minTweak.  This appears to be working correctly on
all parts.